make test directory structure reflect workflow directory structure. move repeated test code to helper functions
