+++ test/rules/all/data/config/peptide.tsv
