+++ test/rules/convert_peptides_txt_to_fasta/data/results/peptides/simulated_peptides_10.txt
