+++ test/rules/convert_peptides_txt_to_fasta/expected/results/fastas/simulated_peptides_10.fasta
