+++ test/rules/qc_alignment/build_diamond_tax_id_lineage/data/resources/ncbi/nodes.dmp
