+++ test/rules/qc_alignment/build_diamond_tax_id_lineage/data/results/diamond/foo-bar-taxids.txt
