+++ test/rules/qc_alignment/build_diamond_tax_id_lineage/expected/results/diamond/foo-bar-lineage.tsv
