+++ test/rules/qc_alignment/map_diamond_accessions_to_taxids/data/resources/uniprot/idmapping_selected.tab.gz
