+++ test/rules/qc_alignment/map_diamond_accessions_to_taxids/data/results/diamond/foo-bar-accessions.txt
