+++ test/rules/qc_alignment/map_diamond_accessions_to_taxids/expected/results/diamond/foo-bar-taxids.txt
