+++ test/rules/run_mmseqs2/mmseqs2_top_10/data/results/mmseqs2/sample.tsv
