+++ test/rules/run_mmseqs2/mmseqs2_top_10/test_mmseqs2_top_10.py
@@ -1,0 +1,18 @@
+from test.common import snakemake_run, check_output
+
+
+def test_mmseqs2_top_10(tmpdir, workflow_path, prepared_workdir):
+    target = "results/mmseqs2_top_10/sample.tsv"
+    snakefile = workflow_path / "workflow/rules/run_mmseqs2.smk"
+
+    # Run the test job.
+    snakemake_run(
+        snakefile,
+        target,
+        prepared_workdir.workdir,
+        additional_arguments=[
+        "--config",
+        "query_dbs=foo",]
+    )
+
+    check_output(prepared_workdir, mode="text")
