+++ test/rules/simulate_sample/convert_scientific_names_to_taxonomy_ids/data/resources/human_microbiome_species.txt
