+++ test/rules/simulate_sample/convert_scientific_names_to_taxonomy_ids/expected/results/simulation/human_microbiome_project_taxonomy_ids.txt
