+++ test/rules/simulate_sample/filter_tax_ids_and_build_lineage/expected/results/simulation/human_microbiome_project_taxonomy_ids_in_uniprot.txt
