+++ test/rules/simulate_sample/map_taxids_to_uniprot_accessions/data/resources/uniprot/idmapping_selected.tab.gz
