+++ test/rules/simulate_sample/map_taxids_to_uniprot_accessions/data/results/simulation/human_microbiome_project_taxonomy_ids.txt
