+++ test/rules/simulate_sample/map_taxids_to_uniprot_accessions/expected/results/simulation/tax2accessions.json
