--- test/unit/conftest.py
@@ -1,8 +1,0 @@
-from pathlib import Path
-
-import pytest
-
-
-@pytest.fixture
-def workflow_path() -> Path:
-    return Path(__file__).parent.parent.parent
